# UltraClone_Spec_Bundle/openapi.yaml
openapi: 3.1.0
info:
  title: UltraClone Voice API
  version: 0.1.0
  description: >
    API for Instant Cloning, Text-to-Speech, Voice Library, Batch jobs, and Quality reports.
    Streaming is provided via WebSocket (`wss://.../v1/stream/{voice_id}`) that sends PCM chunks.
servers:
  - url: http://localhost:5071
paths:
  /v1/voices:
    get:
      summary: List voices
      operationId: listVoices
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/VoiceProfile'
    post:
      summary: Add a new voice (Instant Clone)
      operationId: addVoice
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                reference_audio:
                  type: string
                  format: binary
                notes:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceProfile'
  /v1/voices/{voice_id}:
    get:
      summary: Get voice details
      operationId: getVoice
      parameters:
        - name: voice_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceProfile'
...
components:
  schemas:
    VoiceProfile:
      $ref: './schemas/voice_profile.schema.json'
...